+++ pcs/lib/reports.py
@@ -3246,3 +3246,18 @@ def using_known_host_address_for_host(host_name, address):
             "address": address
         }
     )
+
+
+def resource_in_bundle_not_accessible(
+    bundle_id, inner_resource_id,
+    severity=ReportItemSeverity.ERROR, forceable=report_codes.FORCE_OPTIONS,
+):
+    return ReportItem(
+        report_codes.RESOURCE_IN_BUNDLE_NOT_ACCESSIBLE,
+        severity,
+        info=dict(
+            bundle_id=bundle_id,
+            inner_resource_id=inner_resource_id,
+        ),
+        forceable=forceable,
+    )