+++ pcs/test/test_resource.py
@@ -4874,9 +4874,10 @@ class BundleCommon(
 
     def fixture_bundle(self, name):
         self.assert_pcs_success(
-            "resource bundle create {0} container docker image=pcs:test".format(
-                name
-            )
+            (
+                "resource bundle create {0} container docker image=pcs:test "
+                "network control-port=1234"
+            ).format(name)
         )
 
 
@@ -4911,6 +4912,7 @@ class BundleDeleteTest(BundleCommon):
                 <resources>
                     <bundle id="B">
                         <docker image="pcs:test" />
+                        <network control-port="1234" />
                     </bundle>
                 </resources>
             """,